Adição de diversas partes lógicas novas de autenticação

+++ src/routes/appointment.routes.ts
@@ -6,8 +6,11 @@ import {
 } from '../infra';
 import { AppointmentService } from '../services';
 import { AppointmentController } from '../controllers';
-import { validatorMiddleware } from '../shared';
-import cors from 'cors';
+import {
+  authenticateTokenMiddleware,
+  checkPermission,
+  validatorMiddleware,
+} from '../shared';
 
 const repository: IAppointmentRepository = new PrismaAppointmentRepository(
   prisma
@@ -18,26 +21,58 @@ const controller = new AppointmentController(service);
 const router = Router();
 
 router
-  .get('', (request, response) => {
-    return controller.getAll(request, response);
-  })
-  .get('/:petID', (request, response) => {
-    return controller.getAllByPetID(request, response);
-  })
+  .get(
+    '',
+    authenticateTokenMiddleware,
+    checkPermission(['view_appointments, all_appointment']),
+    (request, response) => {
+      return controller.getAll(request, response);
+    }
+  )
+  .get(
+    '/:petID',
+    checkPermission(['view_appointment, all_appointment']),
+    authenticateTokenMiddleware,
+    (request, response) => {
+      return controller.getAllByPetID(request, response);
+    }
+  )
   .get('/:id', (request, response) => {
     return controller.getOneByID(request, response);
   })
-  .post('/', validatorMiddleware('AppointmentModel'), (request, response) => {
-    return controller.create(request, response);
-  })
-  .put('/', validatorMiddleware('AppointmentModel'), (request, response) => {
-    return controller.update(request, response);
-  })
-  .put('/:id/status/:status', (request, response) => {
-    return controller.updateStatus(request, response);
-  })
-  .delete('/:id', (request, response) => {
-    return controller.delete(request, response);
-  });
+  .post(
+    '/',
+    authenticateTokenMiddleware,
+    checkPermission(['create_appointment, all_appointment']),
+    validatorMiddleware('AppointmentModel'),
+    (request, response) => {
+      return controller.create(request, response);
+    }
+  )
+  .put(
+    '/',
+    authenticateTokenMiddleware,
+    checkPermission(['update_appointment, all_appointment']),
+    validatorMiddleware('AppointmentModel'),
+    (request, response) => {
+      return controller.update(request, response);
+    }
+  )
+  .put(
+    '/:id/status/:status',
+    authenticateTokenMiddleware,
+    checkPermission(['update_appointment, all_appointment']),
+    (request, response) => {
+      return controller.updateStatus(request, response);
+    }
+  )
+  .delete(
+    '/:id',
+    authenticateTokenMiddleware,
+    checkPermission(['delete_appointment, all_appointment']),
+    (request, response) => {
+      return controller.delete(request, response);
+    }
+  );
 
 export default router;
