+++ src/routes/owner.routes.ts
@@ -2,7 +2,11 @@ import { Router } from 'express';
 import { IOwnerRepository, prisma, PrismaOwnerRepository } from '../infra';
 import { OwnerService } from '../services';
 import { OwnerController } from '../controllers';
-import { validatorMiddleware } from '../shared';
+import {
+  authenticateTokenMiddleware,
+  checkPermission,
+  validatorMiddleware,
+} from '../shared';
 
 const repository: IOwnerRepository = new PrismaOwnerRepository(prisma);
 const service = new OwnerService(repository);
@@ -11,20 +15,47 @@ const controller = new OwnerController(service);
 const router = Router();
 
 router
-  .get('', (request, response) => {
-    return controller.getAll(request, response);
-  })
-  .get('/:id', (request, response) => {
-    return controller.getOneByID(request, response);
-  })
-  .post('/', validatorMiddleware('OwnerModel'), (request, response) => {
-    return controller.create(request, response);
-  })
-  .put('/', validatorMiddleware('OwnerModel'), (request, response) => {
-    return controller.update(request, response);
-  })
-  .delete('/:id', (request, response) => {
-    return controller.delete(request, response);
-  });
+  .get(
+    '',
+    authenticateTokenMiddleware,
+    checkPermission(['view_owners, all_owner']),
+    (request, response) => {
+      return controller.getAll(request, response);
+    }
+  )
+  .get(
+    '/:id',
+    authenticateTokenMiddleware,
+    checkPermission(['view_owner, all_owner']),
+    (request, response) => {
+      return controller.getOneByID(request, response);
+    }
+  )
+  .post(
+    '/',
+    authenticateTokenMiddleware,
+    checkPermission(['create_owner, all_owner']),
+    validatorMiddleware('OwnerModel'),
+    (request, response) => {
+      return controller.create(request, response);
+    }
+  )
+  .put(
+    '/',
+    authenticateTokenMiddleware,
+    checkPermission(['update_owner, all_owner']),
+    validatorMiddleware('OwnerModel'),
+    (request, response) => {
+      return controller.update(request, response);
+    }
+  )
+  .delete(
+    '/:id',
+    checkPermission(['delete_owner, all_owner']),
+    authenticateTokenMiddleware,
+    (request, response) => {
+      return controller.delete(request, response);
+    }
+  );
 
 export default router;
