+++ src/routes/pet.routes.ts
@@ -2,7 +2,11 @@ import { Router } from 'express';
 import { IPetRepository, prisma, PrismaPetRepository } from '../infra';
 import { PetService } from '../services';
 import { PetController } from '../controllers';
-import { validatorMiddleware } from '../shared';
+import {
+  authenticateTokenMiddleware,
+  checkPermission,
+  validatorMiddleware,
+} from '../shared';
 
 const repository: IPetRepository = new PrismaPetRepository(prisma);
 const service = new PetService(repository);
@@ -11,23 +15,55 @@ const controller = new PetController(service);
 const router = Router();
 
 router
-  .get('', (request, response) => {
-    return controller.getAll(request, response);
-  })
-  .get('/:ownerID', (request, response) => {
-    return controller.getOneByID(request, response);
-  })
-  .get('/:id', (request, response) => {
-    return controller.getOneByID(request, response);
-  })
-  .post('/', validatorMiddleware('PetModel'), (request, response) => {
-    return controller.create(request, response);
-  })
-  .put('/', validatorMiddleware('PetModel'), (request, response) => {
-    return controller.update(request, response);
-  })
-  .delete('/:id', (request, response) => {
-    return controller.delete(request, response);
-  });
+  .get(
+    '',
+    authenticateTokenMiddleware,
+    checkPermission(['view_pets, all_pet']),
+    (request, response) => {
+      return controller.getAll(request, response);
+    }
+  )
+  .get(
+    '/:ownerID',
+    authenticateTokenMiddleware,
+    checkPermission(['view_pet, all_pet']),
+    (request, response) => {
+      return controller.getOneByID(request, response);
+    }
+  )
+  .get(
+    '/:id',
+    authenticateTokenMiddleware,
+    checkPermission(['view_pet, all_pet']),
+    (request, response) => {
+      return controller.getOneByID(request, response);
+    }
+  )
+  .post(
+    '/',
+    authenticateTokenMiddleware,
+    checkPermission(['create_pet, all_pet']),
+    validatorMiddleware('PetModel'),
+    (request, response) => {
+      return controller.create(request, response);
+    }
+  )
+  .put(
+    '/',
+    authenticateTokenMiddleware,
+    checkPermission(['update_pet, all_pet']),
+    validatorMiddleware('PetModel'),
+    (request, response) => {
+      return controller.update(request, response);
+    }
+  )
+  .delete(
+    '/:id',
+    authenticateTokenMiddleware,
+    checkPermission(['delete_pet, all_pet']),
+    (request, response) => {
+      return controller.delete(request, response);
+    }
+  );
 
 export default router;
